Move CategoriesTabAdapter to saved package
